add dry run

+++ cdsobs/cli/_delete_dataset.py
@@ -41,6 +41,9 @@ def delete_dataset(
         "to delete all partitions of year 1970: 1970-1-1,1970-12-31",
     ),
     version: str = typer.Option(None, help="Version to delete"),
+    dry_run: bool = typer.Option(
+        False, help="Do nothing but print the entries to be deleted."
+    ),
 ):
     """Permanently delete the given dataset from the catalogue and the storage."""
     confirm = prompt(
@@ -56,31 +59,37 @@ def delete_dataset(
 
     with get_session(init_config.catalogue_db) as catalogue_session:
         deleted_entries = delete_from_catalogue(
-            catalogue_session, dataset, dataset_source, time, version=version
+            catalogue_session,
+            dataset,
+            dataset_source,
+            time,
+            version=version,
+            dry_run=dry_run,
         )
-        s3_client = S3Client.from_config(init_config.s3config)
-        try:
-            delete_from_s3(deleted_entries, s3_client)
-        except (Exception, KeyboardInterrupt):
-            catalogue_rollback(catalogue_session, deleted_entries)
-            raise
-        nd = len(deleted_entries)
-        console.print(
-            f"[bold green] {nd} entries deleted from {dataset}. [/bold green]"
-        )
-        nremaining = catalogue_session.scalar(
-            select(func.count())
-            .select_from(Catalogue)
-            .where(Catalogue.dataset == dataset, Catalogue.version == version)
-        )
-        if nremaining == 0:
-            CadsDatasetVersionRepository(catalogue_session).delete_dataset(
-                dataset, version
-            )
+        if not dry_run:
+            s3_client = S3Client.from_config(init_config.s3config)
+            try:
+                delete_from_s3(deleted_entries, s3_client)
+            except (Exception, KeyboardInterrupt):
+                catalogue_rollback(catalogue_session, deleted_entries)
+                raise
+            nd = len(deleted_entries)
             console.print(
-                f"[bold green] Deleted {dataset} {version} from datasets table as it was "
-                f"left empty. [/bold green]"
+                f"[bold green] {nd} entries deleted from {dataset}. [/bold green]"
             )
+            nremaining = catalogue_session.scalar(
+                select(func.count())
+                .select_from(Catalogue)
+                .where(Catalogue.dataset == dataset, Catalogue.version == version)
+            )
+            if nremaining == 0:
+                CadsDatasetVersionRepository(catalogue_session).delete_dataset(
+                    dataset, version
+                )
+                console.print(
+                    f"[bold green] Deleted {dataset} {version} from datasets table as it was "
+                    f"left empty. [/bold green]"
+                )
 
 
 def delete_from_catalogue(
@@ -89,6 +98,7 @@ def delete_from_catalogue(
     dataset_source: str,
     time: str,
     version: str,
+    dry_run: bool = False,
 ):
     catalogue_repo = CatalogueRepository(catalogue_session)
     versions = [
@@ -108,11 +118,17 @@ def delete_from_catalogue(
     entries = catalogue_repo.get_by_filters(filters)
     if not len(entries):
         console.print(f"[red] No entries for dataset {dataset} found")
-    try:
-        catalogue_session.execute(delete(Catalogue).where(*filters))
-        catalogue_session.commit()
-    except (Exception, KeyboardInterrupt):
-        catalogue_rollback(catalogue_session, entries)
+    if dry_run:
+        assets = [e.asset for e in entries]
+        console.print(
+            f"Would delete {len(entries)} entries with the following assets: {assets}"
+        )
+    else:
+        try:
+            catalogue_session.execute(delete(Catalogue).where(*filters))
+            catalogue_session.commit()
+        except (Exception, KeyboardInterrupt):
+            catalogue_rollback(catalogue_session, entries)
     return entries
